Move mappings into renderer
