+++ src/test/kotlin/imgui/gl/test javafx.kt
@@ -56,8 +56,6 @@ class HelloWorld_jfx {
         ctx = Context()
         ImGui.styleColorsDark()
 
-        setupMappings()
-
         while(!ready.get())
             Thread.sleep(1)
 
@@ -150,32 +148,6 @@ class HelloWorld_jfx {
             } else {
                 doCopy()
             }
-        }
-    }
-
-    private fun setupMappings() {
-        with(io) {
-            keyMap[Key.Tab] = KeyCode.TAB.code
-            keyMap[Key.LeftArrow] = KeyCode.LEFT.code
-            keyMap[Key.RightArrow] = KeyCode.RIGHT.code
-            keyMap[Key.UpArrow] = KeyCode.UP.code
-            keyMap[Key.DownArrow] = KeyCode.DOWN.code
-            keyMap[Key.PageUp] = KeyCode.PAGE_UP.code
-            keyMap[Key.PageDown] = KeyCode.PAGE_DOWN.code
-            keyMap[Key.Home] = KeyCode.HOME.code
-            keyMap[Key.End] = KeyCode.END.code
-            keyMap[Key.Insert] = KeyCode.INSERT.code
-            keyMap[Key.Delete] = KeyCode.DELETE.code
-            keyMap[Key.Backspace] = KeyCode.BACK_SPACE.code
-            keyMap[Key.Space] = KeyCode.SPACE.code
-            keyMap[Key.Enter] = KeyCode.ENTER.code
-            keyMap[Key.Escape] = KeyCode.ESCAPE.code
-            keyMap[Key.A] = KeyCode.A.code
-            keyMap[Key.C] = KeyCode.C.code
-            keyMap[Key.V] = KeyCode.V.code
-            keyMap[Key.X] = KeyCode.X.code
-            keyMap[Key.Y] = KeyCode.Y.code
-            keyMap[Key.Z] = KeyCode.Z.code
         }
     }
